# fix: use put_records for kinesis due to rate exceeded