--- a/src/aws/osml/model_runner/app_config.py
+++ b/src/aws/osml/model_runner/app_config.py
@@ -61,7 +61,9 @@ class ServiceConfig:
     throttling_retry_timeout: str = os.getenv("THROTTLING_RETRY_TIMEOUT", "10")
 
     # constant configuration
-    kinesis_max_record_size: str = "1048576"
+    kinesis_max_record_per_batch: str = "500"
+    kinesis_max_record_size_batch: str = "5242880"  # 5 MB in bytes
+    kinesis_max_record_size: str = "1048576"  # 1 MB in bytes
     ddb_max_item_size: str = "200000"
     noop_bounds_model_name: str = "NOOP_BOUNDS_MODEL_NAME"
     noop_geom_model_name: str = "NOOP_GEOM_MODEL_NAME"