**Issue #, if available:** n/a

### Notes

This pull request enhances the KinesisSink class by updating the `_flush_stream` method to improve the handling of records sent to the Kinesis stream. The changes ensure compliance with Kinesis limits, specifically allowing a maximum of 500 records per request, while the total request size does not exceed 5 MB.

**Ran the load test for an hour:**

| Metric                       | Before PR               | This PR               |
|------------------------------|-------------------------|-------------------------|
| Total Images Sent            | 112                     | 115                     |
| Total Images In-Progress     | 14                      | 2                       |
| Total Images Processed       | 98                      | 113                     |
| Total Images Succeeded       | 91                      | 113                     |
| Total Images Failed          | 7                       | 0                       |
| Total GB Processed           | 18.63                   | 30.55                   |
| Total Pixels Processed       | 6,667,674,894           | 10,933,475,922          |

The above load test uses a variety of rareplane images of different sizes (ranging from small to large) and runs for one hour, sending an image request every 30 seconds.

### Checklist

Before you submit a pull request, please make sure you have the following:
- [x] Code changes are compact and well-structured to facilitate easy review
- [x] Changes are documented in the README.md and other relevant documentation pages
- [x] PR title and description accurately reflect the changes and are detailed enough for historical tracking
- [x] PR contains tests that cover all new code and the code has been manual tested
- [x] All new dependencies are declared (if any), and no unnecessary libraries are added
- [x] Performance impacts (if any) of the changes are evaluated and documented
- [x] Security implications of the changes (if any) are reviewed and addressed
- [x] I have read the [Contributing Guidelines](../CONTRIBUTING.md) and agree to follow the [Code of Conduct](../CODE_OF_CONDUCT.md)

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.